requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.20
boto3>=1.34.0
botocore>=1.34.0

# Vector stores
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Document processing
PyPDF2>=3.0.1
pypdf>=4.0.0
unstructured>=0.12.0
python-magic>=0.4.27

# Docling for advanced PDF and image processing
docling>=1.0.0
docling-core>=1.0.0
docling-ibm-models>=1.0.0

# Multimodal processing (optional)
pillow>=10.0.0
pymupdf>=1.23.0
opencv-python>=4.8.0

# LangGraph for agentic workflows (optional)
langgraph>=0.0.30

# Web UI
gradio>=4.0.0

# Configuration and utilities
pyyaml>=6.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.0.0

# Text processing
tiktoken>=0.5.0
sentence-transformers>=2.2.0

# Optional: For enhanced PDF processing
#unstructured[pdf]>=0.12.0
#detectron2  # For advanced layout detection

# Development dependencies (uncomment if needed)
#pytest>=7.0.0
#black>=23.0.0
#flake8>=6.0.0
#mypy>=1.0.0
